fixed fab button to slide up when snack bar appears
