--- app/src/main/res/layout/button_sub_menu_1.xml
@@ -1,123 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<android.support.constraint.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    android:id="@+id/button_constraint_layout"
-    android:layout_width="wrap_content"
-    android:layout_height="wrap_content">
-
-    <android.support.design.widget.FloatingActionButton
-        android:id="@+id/add_button"
-        android:layout_width="64dp"
-        android:layout_height="64dp"
-        android:layout_gravity="bottom|end"
-        android:layout_marginBottom="8dp"
-        android:layout_marginEnd="8dp"
-        android:layout_marginStart="8dp"
-        android:scaleType="centerCrop"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintHorizontal_bias="1.0"
-        app:layout_constraintStart_toStartOf="parent"
-        app:srcCompat="@drawable/ic_add_black_24dp" />
-
-    <android.support.design.widget.FloatingActionButton
-        android:id="@+id/add_simple_note_button"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginBottom="8dp"
-        android:clickable="true"
-        android:focusable="true"
-        app:layout_constraintBottom_toTopOf="@+id/add_button"
-        app:layout_constraintEnd_toStartOf="@+id/add_checklist_button"
-        app:layout_constraintStart_toEndOf="@+id/add_checklist_button"
-        app:srcCompat="@drawable/ic_note_add_black_24dp" />
-
-    <android.support.design.widget.FloatingActionButton
-        android:id="@+id/add_checklist_button"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginBottom="8dp"
-        android:layout_marginEnd="8dp"
-        android:layout_marginStart="8dp"
-        android:clickable="true"
-        android:focusable="true"
-        app:layout_constraintBottom_toTopOf="@+id/add_simple_note_button"
-        app:layout_constraintEnd_toStartOf="@+id/add_image_note_button"
-        app:layout_constraintStart_toEndOf="@+id/add_image_note_button"
-        app:srcCompat="@drawable/ic_format_list_bulleted_black_24dp" />
-
-    <android.support.design.widget.FloatingActionButton
-        android:id="@+id/add_image_note_button"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginBottom="8dp"
-        android:layout_marginLeft="8dp"
-        android:layout_marginStart="8dp"
-        android:layout_marginTop="8dp"
-        android:clickable="true"
-        android:focusable="true"
-        app:layout_constraintBottom_toTopOf="@+id/add_checklist_button"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintHorizontal_bias="0.93"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent"
-        app:layout_constraintVertical_bias="1.0"
-        app:srcCompat="@drawable/ic_insert_photo_black_24dp" />
-
-    <TextView
-        android:id="@+id/tv_add_simple_note"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginEnd="8dp"
-        android:layout_marginStart="8dp"
-        android:background="@android:color/white"
-        android:elevation="4dp"
-        android:padding="5dp"
-        android:text="@string/simple_note"
-        android:textAppearance="@style/TextAppearance.AppCompat.Subhead"
-        app:layout_constraintBottom_toTopOf="@+id/add_simple_note_button"
-        app:layout_constraintEnd_toStartOf="@+id/add_simple_note_button"
-        app:layout_constraintHorizontal_bias="0.9"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/add_simple_note_button" />
-
-    <TextView
-        android:id="@+id/tv_add_checklist"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginLeft="8dp"
-        android:layout_marginStart="8dp"
-        android:background="@android:color/white"
-        android:elevation="4dp"
-        android:padding="5dp"
-        android:text="@string/checklist"
-        android:textAppearance="@style/TextAppearance.AppCompat.Subhead"
-        app:layout_constraintBottom_toTopOf="@+id/add_checklist_button"
-        app:layout_constraintEnd_toStartOf="@+id/add_checklist_button"
-        app:layout_constraintStart_toStartOf="@+id/tv_add_simple_note"
-        app:layout_constraintTop_toBottomOf="@+id/add_checklist_button" />
-
-    <TextView
-        android:id="@+id/tv_add_image_note"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginEnd="8dp"
-        android:layout_marginStart="8dp"
-        android:background="@android:color/white"
-        android:elevation="4dp"
-        android:padding="5dp"
-        android:text="@string/image_note"
-        android:textAppearance="@style/TextAppearance.AppCompat.Subhead"
-        app:layout_constraintBottom_toTopOf="@+id/add_image_note_button"
-        app:layout_constraintEnd_toStartOf="@+id/add_image_note_button"
-        app:layout_constraintStart_toStartOf="@+id/tv_add_simple_note"
-        app:layout_constraintTop_toBottomOf="@+id/add_image_note_button" />
-
-    <android.support.constraint.Group
-        android:id="@+id/button_sub_menu"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        app:constraint_referenced_ids="add_simple_note_button,tv_add_image_note
-,tv_add_checklist,tv_add_simple_note,add_image_note_button,add_checklist_button" />
-
-</android.support.constraint.ConstraintLayout>
