+++ app/src/main/res/layout/home_page_layout.xml
@@ -22,14 +22,14 @@
             android:layout_width="match_parent"
             android:layout_height="wrap_content" />
     </LinearLayout>
-    <include
+    <!--<include
         layout="@layout/button_sub_menu"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:layout_alignParentEnd="true"
         android:layout_alignParentBottom="true"
         android:layout_margin="8dp"
-        android:layout_alignParentRight="true" />
+        android:layout_alignParentRight="true" />-->
     <ProgressBar
         android:id="@+id/progress_bar"
         android:layout_width="56dp"
